+++ src/main/java/com/springboot/ecommerce/service/UserService.java
@@ -1,42 +1,65 @@
 package com.springboot.ecommerce.service;
 
+import com.springboot.ecommerce.entity.Product;
 import com.springboot.ecommerce.entity.User;
 import com.springboot.ecommerce.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class UserService  {
 
     private final UserRepository userRepo;
-    public User findById(Long id) {
-        return userRepo.findById(id).orElse(null);
+    public ResponseEntity<?> findById(Long id) {
+        Optional<User> user= userRepo.findById(id);
+        if (user.isEmpty())
+            return ResponseEntity.badRequest().body("This User doesn't exist!");
+
+        return ResponseEntity.ok(user);
     }
 
-    public User findByName(String name) {
-        return userRepo.findByName(name);
+    public ResponseEntity<?> findByName(String name) {
+        User user= userRepo.findByName(name);
+        if (user== null)
+            return ResponseEntity.badRequest().body("Product with this name doesn't exist");
+        return ResponseEntity.ok(user);    }
+
+    public ResponseEntity<?> findByEmail(String email) {
+        User user= userRepo.findByEmail(email);
+        if (user== null)
+            return ResponseEntity.badRequest().body("User with this email doesn't exist");
+        return ResponseEntity.ok(user);
     }
 
-    public User findByEmail(String email) {
-        return userRepo.findByEmail(email);
-    }
-
-    public List<User> findAll() {
-        return userRepo.findAll();
+    public ResponseEntity<?> findAll() {
+        List<User> users= userRepo.findAll();
+        if (users.isEmpty())
+            return ResponseEntity.badRequest().body("There are no users");
+        return ResponseEntity.ok(users);
     }
 
     public User insert(User user) {
         return userRepo.save(user);
     }
 
-    public User update(User user) {
-        return userRepo.save(user);
+    public ResponseEntity<?> update(User user) {
+        Long id= user.getId();
+        Optional<User> temp= userRepo.findById(id);
+        if (temp.isEmpty())
+            return ResponseEntity.badRequest().body("This user doesn't exist!");
+        return ResponseEntity.ok(userRepo.save(user));
     }
 
-    public void delete(Long id) {
+    public ResponseEntity<?> delete(Long id) {
+        User user= userRepo.findById(id).get();
+        if (user== null)
+            return ResponseEntity.badRequest().body("This user doesn't exist");
         userRepo.deleteById(id);
+        return ResponseEntity.ok("User with id "+ id +" has been deleted successfully!");
     }
 }
